requirements.txt
replicate~=0.26.0
streamlit~=1.35.0
streamlit_image_select
python-dotenv
